Switch to AggregatingMergeTree and record insert times
The original choice of `ReplacingMergeTree` table engine worked fine,
but it's limiting when it comes to adding more data. The point of
using a non-default table engine is to use ClickHouse to deduplicate
inserted data. But there's more flexible table engines available.

This commit switches to AggregatingMergeTree and adds a new `inserted`
column that will expose the time a reading was first inserted into
ClickHouse. This will help with understanding when the Octopus API
updates and when is best to query it.